Close server sockets in plugin deinit callback
If we don't do this, the ports may remain marked as in use after the plugin
is deinitialized, and on next plugin load, for example during server restart,
it would almost certainly fail to bind to them (at least on Linux).

Also, set the SO_REUSEADDR flag to avoid the delay between server restarts.